Enable metal
This enables using metal by adding an additional target for ggml-metal
and then optionally including it in the build.

I also changed `ifdef SWIFT_PACKAGE` to `ifdef SWIFTPM_MODULE_BUNDLE`
since I had the former variable set and not the latter (not sure why).

This works on my machine, but I'm not a rapper, so please let me know if
I made a mistake.